+++ sonar-plugin/src/main/java/io/ecocode/javascript/checks/AvoidHighAccuracyGeolocation.java
@@ -26,7 +26,7 @@ import org.sonarsource.analyzer.commons.annotations.DeprecatedRuleKey;
 @JavaScriptRule
 @TypeScriptRule
 @Rule(key = AvoidHighAccuracyGeolocation.RULE_KEY)
-@DeprecatedRuleKey(repositoryKey = "ecocode-javascript", ruleKey = "EC8")
+@DeprecatedRuleKey(ruleKey = "EC8")
 public class AvoidHighAccuracyGeolocation implements EslintBasedCheck {
 
     public static final String RULE_KEY = "EC523";